pyproject.toml
[tool.poetry]
name = "django_ninja_auth"
version = "1.0.0"
description = "Django authorization views adapted to django-ninja"
license = "WTFPL"
readme = "README.md"
homepage = "https://github.com/mugartec/django-ninja-auth"
repository = "https://github.com/mugartec/django-ninja-auth"
keywords = ["django", "django-ninja", "auth"]
packages = [{include = "ninja_auth"}]
authors = ["Martín Ugarte <contact@martinugarte.com>"]
include = [
    "LICENSE", {path = "tests", format="sdist"}
]

[tool.poetry.dependencies]
python = "^3.8"
django-ninja = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
